+++ CHANGES.rst
@@ -6,6 +6,12 @@ JWST
 
 -  update niriss pars-jumpstep parkeys [#890]
 
+General
+-------
+
+- Revised core.utils to allow I/O to work under Windows [#891]
+
+
 11.16.5 (2022-06-27)
 ====================